move commands to one file to clear up clutter in the root directory
